move all tests related to the `SourceSink` conversion to `test_converters_source_sink.py`
