+++ tests/tools/test_converters_source_sink.py
@@ -1,11 +1,13 @@
 from pathlib import Path
-from unittest.mock import patch
+from typing import Dict
+from unittest.mock import Mock, patch
 
 import pytest
 
 from hydrolib.core.dflowfm.ext.models import SourceSink
 from hydrolib.core.dflowfm.extold.models import ExtOldForcing, ExtOldQuantity
 from hydrolib.tools.extforce_convert.converters import SourceSinkConverter
+from hydrolib.tools.extforce_convert.main_converter import ExternalForcingConverter
 
 
 class TestParseTimFileForSourceSink:
@@ -405,3 +407,81 @@ class TestSourceSinkConverter:
         assert data[1].loc[:, 0].to_list() == [4.0, 4.0, 4.0, 4.0, 4.0]
         # discharge
         assert data[0].loc[:, 0].to_list() == [1.0, 1.0, 1.0, 1.0, 1.0]
+
+
+class TestMainConverter:
+
+    def test_sources_sinks_only(self, old_forcing_file_boundary: Dict[str, str]):
+        """
+        The old external forcing file contains only 3 quantities `discharge_salinity_temperature_sorsin`,
+        `initialsalinity`, and `initialtemperature`.
+
+        - polyline 2*3 file `leftsor.pliz` is used to read the source and sink points.
+        - tim file `tim-3-columns.tim` with 3 columns (plus the time column) the name should be the same as the
+        polyline but the `tim-3-columns.tim` is mocked in the test.
+
+        """
+        path = "tests/data/input/source-sink/source-sink.ext"
+        mdu_info = {
+            "refdate": "minutes since 2015-01-01 00:00:00",
+        }
+        converter = ExternalForcingConverter(path, mdu_info=mdu_info)
+        # Mock the fm_model
+        mock_fm_model = Mock()
+        converter._fm_model = mock_fm_model
+
+        tim_file = Path("tim-3-columns.tim")
+        with patch("pathlib.Path.with_suffix", return_value=tim_file):
+            ext_model, inifield_model, structure_model = converter.update()
+
+        # all the quantities in the old external file are initial conditions
+        # check that all the quantities (3) were converted to initial conditions
+        num_quantities = 1
+        assert len(ext_model.sourcesink) == num_quantities
+        # no parameters or any other structures, lateral or meteo data
+        assert len(inifield_model.parameter) == 0
+        assert len(ext_model.lateral) == 0
+        assert len(ext_model.meteo) == 0
+        assert len(structure_model.structure) == 0
+        assert len(inifield_model.initial) == 2
+        quantities = ext_model.sourcesink
+        quantities[0].name = "discharge_salinity_temperature_sorsin"
+
+    def test_sources_sinks_with_fm(self, old_forcing_file_boundary: Dict[str, str]):
+        """
+        The old external forcing file contains only 3 quantities `discharge_salinity_temperature_sorsin`,
+        `initialsalinity`, and `initialtemperature`.
+
+        - polyline 2*3 file `leftsor.pliz` is used to read the source and sink points.
+        - tim file `tim-3-columns.tim` with 3 columns (plus the time column) the name should be the same as the
+        polyline but the `tim-3-columns.tim` is mocked in the test.
+
+        """
+        path = "tests/data/input/source-sink/source-sink.ext"
+        mdu_info = {
+            "refdate": "minutes since 2015-01-01 00:00:00",
+            "salinity": True,
+            "temperature": True,
+        }
+        converter = ExternalForcingConverter(path, mdu_info=mdu_info)
+
+        # Mock the fm_model
+        mock_fm_model = Mock()
+        converter._fm_model = mock_fm_model
+
+        tim_file = Path("tim-3-columns.tim")
+        with patch("pathlib.Path.with_suffix", return_value=tim_file):
+            ext_model, inifield_model, structure_model = converter.update()
+
+        # all the quantities in the old external file are initial conditions
+        # check that all the quantities (3) were converted to initial conditions
+        num_quantities = 1
+        assert len(ext_model.sourcesink) == num_quantities
+        # no parameters or any other structures, lateral or meteo data
+        assert len(inifield_model.parameter) == 0
+        assert len(ext_model.lateral) == 0
+        assert len(ext_model.meteo) == 0
+        assert len(structure_model.structure) == 0
+        assert len(inifield_model.initial) == 2
+        quantities = ext_model.sourcesink
+        quantities[0].name = "discharge_salinity_temperature_sorsin"
